test: Properly test `trap` functions
Previously, the `core.trap_*` functions were not fully tested because Bats
appears to be intercepting the signals (they were only tested by testing
for its side effects, like saving global variables). This was fixed by running
everything in a subprocess. Since the Bash builtin helpers are not
available in this environment, some things like `load` had to be renamed
to `source`, or removed entirely.