breaking: Deprecate `print_die` functions
These functions were initially used because a 'die' function is
necessary in most scripts. Because of othe other `_fn`s, a
`print_die_fn` was made, which makes little sense, as if the function
name is required to be printed on fault, it would be better to show the
stacktrace anyways (and implement some functionality for that)

There was also an issue of switching the `die` function from calling
`core.print_error` to `core.print_fatal`. This semantically causes some
issues; for example, internal faults should be treated as "fatal", while
user-input errors should be treated as "errors".

Any `die` functions should be created at hyperupcall/bash-std